Use one-way binding to avoid cell value modification before validation
The editor's input value is updated when the cell value changes, but
the cell value is not updated when the editor's input value changes.
This allows to validate the editor value first, and then to update the
cell value once validation completes, in two separate steps. Before it
was updated, and then reverted if validation failed.