Do not modify a destroyed cell after validation
When deleting a row/column during a cell validation (easy when there are
action buttons), then we will call set on a destroyed object and Ember
does not like it and displays an error.

This commit adds a check before modifying a cell after validation. If
the cell has been destroyed, then it exits.